refactor: split latency checker runtime
